# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORBIT Autonomous Hiring OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SkillScore {
  skill: string;
  score: number;
  category: string;
}

export interface Candidate {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  jobId?: number | null;
  stage: string;
  overallScore: number;
  technicalScore: number;
  cultureScore: number;
  growthScore: number;
  retentionProbability: number;
  hiddenTalent: boolean;
  /** @nullable */
  verdict?: string | null;
  avatarInitials: string;
  location: string;
  yearsExperience: number;
  currentCompany: string;
  skills: SkillScore[];
  summary: string;
  githubActivity: string;
  createdAt: string;
}

export interface CandidateInput {
  name: string;
  email: string;
  role: string;
  /** @nullable */
  jobId?: number | null;
  location?: string;
  yearsExperience?: number;
  currentCompany?: string;
  summary?: string;
}

export interface CandidateUpdate {
  name?: string;
  stage?: string;
  verdict?: string;
}

export interface CandidateTwin {
  candidateId: number;
  skillDna: SkillScore[];
  learningVelocity: number;
  leadershipPotential: number;
  adaptabilityScore: number;
  communicationScore: number;
  careerGoals: string;
  growthTrajectory: string;
  strengths: string[];
  weaknesses: string[];
  futureSuccessPrediction: number;
  retentionForecast: number;
  promotionTimeline: string;
}

export interface PerformancePoint {
  month: string;
  performance: number;
  productivity: number;
  teamImpact: number;
}

export interface PerformanceSimulation {
  candidateId: number;
  threeMonthScore: number;
  sixMonthScore: number;
  twelveMonthScore: number;
  promotionPotential: number;
  leadershipDevelopment: number;
  teamInfluence: number;
  learningProgression: number;
  forecast: PerformancePoint[];
  summary: string;
}

export interface Job {
  id: number;
  title: string;
  department: string;
  location: string;
  description: string;
  requirements: string[];
  candidateCount: number;
  status: string;
  createdAt: string;
}

export interface JobInput {
  title: string;
  department: string;
  location: string;
  description: string;
  requirements?: string[];
}

export interface JobUpdate {
  title?: string;
  status?: string;
  description?: string;
}

export interface PipelineStageSummary {
  stage: string;
  label: string;
  count: number;
  candidates: Candidate[];
}

export interface PipelineSummary {
  stages: PipelineStageSummary[];
  totalCandidates: number;
  activeJobs: number;
}

export interface PipelineMove {
  candidateId: number;
  stage: string;
}

export interface DebateMessage {
  agent: string;
  agentType: string;
  message: string;
  stance: string;
  evidence: string[];
  timestamp: string;
}

export interface Debate {
  id: number;
  candidateId: number;
  candidateName: string;
  jobTitle: string;
  messages: DebateMessage[];
  verdict: string;
  confidenceScore: number;
  technicalScore: number;
  cultureScore: number;
  growthScore: number;
  riskScore: number;
  summary: string;
  status: string;
  createdAt: string;
}

export interface DebateInput {
  candidateId: number;
  /** @nullable */
  jobId?: number | null;
}

export interface TeamMember {
  name: string;
  role: string;
  skills: string[];
  leadershipStyle: string;
}

export interface ChemistryDimension {
  dimension: string;
  score: number;
  description: string;
}

export interface TeamChemistry {
  overallScore: number;
  skillComplementarity: number;
  leadershipBalance: number;
  communicationCompatibility: number;
  teamSynergy: number;
  collaborationPotential: number;
  dimensions: ChemistryDimension[];
  members: TeamMember[];
  insights: string[];
}

export interface TeamChemistryPreview {
  currentScore: number;
  projectedScore: number;
  delta: number;
  improvements: string[];
  risks: string[];
  recommendation: string;
}

export interface ChemistryPreviewInput {
  candidateId: number;
}

export interface MissionStep {
  id: string;
  label: string;
  status: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  details?: string | null;
}

export interface Mission {
  id: number;
  /** @nullable */
  jobId?: number | null;
  jobTitle: string;
  status: string;
  steps: MissionStep[];
  candidatesDiscovered: number;
  hiddenTalentsFound: number;
  debatesRun: number;
  /** @nullable */
  topCandidateId?: number | null;
  /** @nullable */
  topCandidateName?: string | null;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface MissionInput {
  /** @nullable */
  jobId?: number | null;
  jobTitle: string;
}

export type DashboardAnalyticsHiringVelocityItem = {
  week: string;
  candidates: number;
  offers: number;
};

export interface AgentActivity {
  id: string;
  agent: string;
  agentType: string;
  action: string;
  target: string;
  timestamp: string;
  status: string;
}

export interface DashboardAnalytics {
  totalCandidates: number;
  activeJobs: number;
  hiddenTalentsFound: number;
  activeAgents: number;
  debatesCompleted: number;
  avgMatchScore: number;
  timeToHireReduction: number;
  topCandidate?: Candidate;
  recentActivity: AgentActivity[];
  hiringVelocity: DashboardAnalyticsHiringVelocityItem[];
}

export interface FunnelStage {
  stage: string;
  count: number;
  conversionRate: number;
}

export type ListCandidatesParams = {
/**
 * @nullable
 */
jobId?: number | null;
/**
 * @nullable
 */
stage?: string | null;
/**
 * @nullable
 */
search?: string | null;
};