# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ORBIT Autonomous Hiring OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - in: query
          name: jobId
          schema:
            type: ["integer", "null"]
        - in: query
          name: stage
          schema:
            type: ["string", "null"]
        - in: query
          name: search
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /candidates/{id}/twin:
    get:
      operationId: getCandidateTwin
      tags: [candidates]
      summary: Get digital twin for a candidate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Digital twin data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateTwin"
  /candidates/{id}/simulation:
    get:
      operationId: getCandidateSimulation
      tags: [candidates]
      summary: Get performance simulation for a candidate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Simulation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceSimulation"
...
  /autopilot/missions:
    get:
      operationId: listMissions
      tags: [autopilot]
      summary: List all autopilot missions
      responses:
        "200":
          description: Missions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mission"
    post:
      operationId: launchMission
      tags: [autopilot]
      summary: Launch an autonomous hiring mission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionInput"
      responses:
        "201":
          description: Mission launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /autopilot/missions/{id}:
    get:
      operationId: getMission
      tags: [autopilot]
      summary: Get a mission's status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
...
components:
  schemas:
...
    SkillScore:
      type: object
      properties:
        skill:
          type: string
        score:
          type: number
        category:
          type: string
      required: [skill, score, category]
    Candidate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        jobId:
          type: ["integer", "null"]
        stage:
          type: string
        overallScore:
          type: number
        technicalScore:
          type: number
        cultureScore:
          type: number
        growthScore:
          type: number
        retentionProbability:
          type: number
        hiddenTalent:
          type: boolean
        verdict:
          type: ["string", "null"]
        avatarInitials:
          type: string
        location:
          type: string
        yearsExperience:
          type: number
        currentCompany:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillScore"
        summary:
          type: string
        githubActivity:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - stage
        - overallScore
        - technicalScore
        - cultureScore
        - growthScore
        - retentionProbability
        - hiddenTalent
        - avatarInitials
        - location
        - yearsExperience
        - currentCompany
        - skills
        - summary
        - githubActivity
        - createdAt
    CandidateInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        jobId:
          type: ["integer", "null"]
        location:
          type: string
        yearsExperience:
          type: number
        currentCompany:
          type: string
        summary:
          type: string
      required: [name, email, role]
    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        stage:
          type: string
        verdict:
          type: string
    CandidateTwin:
      type: object
      properties:
        candidateId:
          type: integer
        skillDna:
          type: array
          items:
            $ref: "#/components/schemas/SkillScore"
        learningVelocity:
          type: number
        leadershipPotential:
          type: number
        adaptabilityScore:
          type: number
        communicationScore:
          type: number
        careerGoals:
          type: string
        growthTrajectory:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        futureSuccessPrediction:
          type: number
        retentionForecast:
          type: number
        promotionTimeline:
          type: string
      required:
        - candidateId
        - skillDna
        - learningVelocity
        - leadershipPotential
        - adaptabilityScore
        - communicationScore
        - careerGoals
        - growthTrajectory
        - strengths
        - weaknesses
        - futureSuccessPrediction
        - retentionForecast
        - promotionTimeline
    PerformancePoint:
      type: object
      properties:
        month:
          type: string
        performance:
          type: number
        productivity:
          type: number
        teamImpact:
          type: number
      required: [month, performance, productivity, teamImpact]
    PerformanceSimulation:
      type: object
      properties:
        candidateId:
          type: integer
        threeMonthScore:
          type: number
        sixMonthScore:
          type: number
        twelveMonthScore:
          type: number
        promotionPotential:
          type: number
        leadershipDevelopment:
          type: number
        teamInfluence:
          type: number
        learningProgression:
          type: number
        forecast:
          type: array
          items:
            $ref: "#/components/schemas/PerformancePoint"
        summary:
          type: string
      required:
        - candidateId
        - threeMonthScore
        - sixMonthScore
        - twelveMonthScore
        - promotionPotential
        - leadershipDevelopment
        - teamInfluence
        - learningProgression
        - forecast
        - summary
...
    MissionStep:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        status:
          type: string
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
      required: [id, label, status]
    Mission:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        jobTitle:
          type: string
        status:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/MissionStep"
        candidatesDiscovered:
          type: integer
        hiddenTalentsFound:
          type: integer
        debatesRun:
          type: integer
        topCandidateId:
          type: ["integer", "null"]
        topCandidateName:
          type: ["string", "null"]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - jobTitle
        - status
        - steps
        - candidatesDiscovered
        - hiddenTalentsFound
        - debatesRun
        - startedAt
    MissionInput:
      type: object
      properties:
        jobId:
          type: ["integer", "null"]
        jobTitle:
          type: string
      required: [jobTitle]
...