# lib/api-zod/src/generated/types/missionStep.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORBIT Autonomous Hiring OS API
 * OpenAPI spec version: 0.1.0
 */

export interface MissionStep {
  id: string;
  label: string;
  status: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  details?: string | null;
}